% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ggplot-add.R
\name{add_axis_title_y}
\alias{add_axis_title_y}
\title{Add axis title y with styling}
\usage{
add_axis_title_y(
  label = NULL,
  size = NULL,
  face = NULL,
  color = NULL,
  family = NULL
)
}
\arguments{
\item{label}{Y-axis title text. Default NULL.}

\item{size}{Font size. Default NULL.}

\item{face}{Font face. Default NULL.}

\item{color}{Font color. Default NULL.}

\item{family}{Font family. Default NULL.}
}
\value{
A list of ggplot2 components (labs + theme).
}
\description{
Add axis title y with styling
}
\examples{
library(ggplot2)
p <- ggplot(mtcars, aes(wt, mpg)) + geom_point()
p + add_axis_title_y("MPG", size = 14, color = "darkgreen")
p + add_axis_title_y("Y Label")
}
