% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ggplot-add.R
\name{add_caption}
\alias{add_caption}
\title{Add caption with styling}
\usage{
add_caption(
  label = NULL,
  size = NULL,
  face = NULL,
  color = NULL,
  hjust = NULL,
  family = NULL
)
}
\arguments{
\item{label}{Caption text. Default NULL.}

\item{size}{Font size. Default NULL.}

\item{face}{Font face. Default NULL.}

\item{color}{Font color. Default NULL.}

\item{hjust}{Horizontal justification. Default NULL.}

\item{family}{Font family. Default NULL.}
}
\value{
A list of ggplot2 components (labs + theme).
}
\description{
Add caption with styling
}
\examples{
library(ggplot2)
p <- ggplot(mtcars, aes(wt, mpg)) + geom_point()
p + add_caption("Data: mtcars", size = 8, color = "grey60")
p + add_caption("Right aligned", hjust = 1)
}
