% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ggplot-add.R
\name{add_legend_title}
\alias{add_legend_title}
\title{Add legend title with styling}
\usage{
add_legend_title(
  label = NULL,
  size = NULL,
  face = NULL,
  color = NULL,
  family = NULL
)
}
\arguments{
\item{label}{Legend title text. Default NULL.}

\item{size}{Font size. Default NULL.}

\item{face}{Font face. Default NULL.}

\item{color}{Font color. Default NULL.}

\item{family}{Font family. Default NULL.}
}
\value{
A list of ggplot2 components (labs + theme).
}
\description{
Add legend title with styling
}
\examples{
library(ggplot2)
p <- ggplot(mtcars, aes(wt, mpg, color = factor(cyl))) + geom_point()
p + add_legend_title("Cylinders", size = 12, color = "steelblue")
p + add_legend_title("No. of Cyl")
}
