% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ggplot-add.R
\name{add_tag}
\alias{add_tag}
\title{Add tag with styling}
\usage{
add_tag(
  label = NULL,
  size = NULL,
  face = NULL,
  color = NULL,
  position = NULL,
  family = NULL
)
}
\arguments{
\item{label}{Tag text. Default NULL.}

\item{size}{Font size. Default NULL.}

\item{face}{Font face. Default NULL.}

\item{color}{Font color. Default NULL.}

\item{position}{Tag position. Default NULL.}

\item{family}{Font family. Default NULL.}
}
\value{
A list of ggplot2 components (labs + theme).
}
\description{
Add tag with styling
}
\examples{
library(ggplot2)
p <- ggplot(mtcars, aes(wt, mpg)) + geom_point()
p + add_tag("A", size = 12, color = "red")
p + add_tag("(a)", position = "bottomright")
}
