% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ggplot-adjust.R
\name{adjust_axis_line}
\alias{adjust_axis_line}
\title{Adjust axis lines appearance}
\usage{
adjust_axis_line(color = NULL, linewidth = NULL, linetype = NULL)
}
\arguments{
\item{color}{Line color. NULL to keep default.}

\item{linewidth}{Line width. Numeric. NULL to keep default.}

\item{linetype}{Line type. "solid", "dashed", etc. NULL to keep default.}
}
\value{
A ggplot2 theme object.
}
\description{
Adjust axis lines appearance
}
\examples{
library(ggplot2)
p <- ggplot(mtcars, aes(wt, mpg)) + geom_point()

# Gray axis lines
p + adjust_axis_line(color = "#888888", linewidth = 0.5)

# Dashed axis lines
p + adjust_axis_line(linetype = "dashed")
}
