% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ggplot-adjust.R
\name{adjust_axis_line_y}
\alias{adjust_axis_line_y}
\title{Adjust y-axis line appearance}
\usage{
adjust_axis_line_y(color = NULL, linewidth = NULL, linetype = NULL)
}
\arguments{
\item{color}{Line color. NULL to keep default.}

\item{linewidth}{Line width. Numeric. NULL to keep default.}

\item{linetype}{Line type. "solid", "dashed", etc. NULL to keep default.}
}
\value{
A ggplot2 theme object.
}
\description{
Adjust y-axis line appearance
}
\examples{
library(ggplot2)
p <- ggplot(mtcars, aes(wt, mpg)) + geom_point()

# Only show y-axis line
p + adjust_axis_line_y(color = "#333333", linewidth = 0.8)
}
