% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ggplot-adjust.R
\name{adjust_axis_ticks}
\alias{adjust_axis_ticks}
\title{Adjust axis ticks appearance}
\usage{
adjust_axis_ticks(color = NULL, linewidth = NULL, linetype = NULL)
}
\arguments{
\item{color}{Line color. NULL to keep default.}

\item{linewidth}{Line width. Numeric. NULL to keep default.}

\item{linetype}{Line type. "solid", "dashed", etc. NULL to keep default.}
}
\value{
A ggplot2 theme object.
}
\description{
Adjust axis ticks appearance
}
\examples{
library(ggplot2)
p <- ggplot(mtcars, aes(wt, mpg)) + geom_point()

# Thinner ticks
p + adjust_axis_ticks(linewidth = 0.3, color = "#888888")
}
