% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ggplot-adjust.R
\name{adjust_axis_ticks_x}
\alias{adjust_axis_ticks_x}
\title{Adjust x-axis ticks appearance}
\usage{
adjust_axis_ticks_x(color = NULL, linewidth = NULL, linetype = NULL)
}
\arguments{
\item{color}{Line color. NULL to keep default.}

\item{linewidth}{Line width. Numeric. NULL to keep default.}

\item{linetype}{Line type. "solid", "dashed", etc. NULL to keep default.}
}
\value{
A ggplot2 theme object.
}
\description{
Adjust x-axis ticks appearance
}
\examples{
library(ggplot2)
p <- ggplot(mtcars, aes(wt, mpg)) + geom_point()

# Custom x-axis ticks
p + adjust_axis_ticks_x(color = "#2166AC", linewidth = 0.5)
}
