% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ggplot-adjust.R
\name{adjust_family}
\alias{adjust_family}
\title{Can be added directly to a ggplot with \code{+} operator.}
\usage{
adjust_family(family = "sans")
}
\arguments{
\item{family}{Font family name. e.g., "serif", "sans", "Arial", "Times New Roman"}
}
\value{
A ggplot2 theme object
}
\description{
Can be added directly to a ggplot with \code{+} operator.
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(x = wt, y = mpg)) +
  geom_point() +
  adjust_family("serif")
}
