% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ggplot-adjust.R
\name{adjust_panel_grid_major_x}
\alias{adjust_panel_grid_major_x}
\title{Adjust x-direction major grid lines appearance}
\usage{
adjust_panel_grid_major_x(color = NULL, linewidth = NULL, linetype = NULL)
}
\arguments{
\item{color}{Line color. NULL to keep default.}

\item{linewidth}{Line width. NULL to keep default.}

\item{linetype}{Line type. NULL to keep default.}
}
\value{
A ggplot2 theme object.
}
\description{
Adjust x-direction major grid lines appearance
}
\examples{
library(ggplot2)
p <- ggplot(mtcars, aes(wt, mpg)) + geom_point()

# Custom x-direction grid
p + adjust_panel_grid_major_x(color = "#EEEEEE", linewidth = 0.3)
}
