% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ggplot-adjust.R
\name{adjust_panel_grid_minor}
\alias{adjust_panel_grid_minor}
\title{Adjust minor grid lines appearance}
\usage{
adjust_panel_grid_minor(color = NULL, linewidth = NULL, linetype = NULL)
}
\arguments{
\item{color}{Line color. NULL to keep default.}

\item{linewidth}{Line width. NULL to keep default.}

\item{linetype}{Line type. NULL to keep default.}
}
\value{
A ggplot2 theme object.
}
\description{
Adjust minor grid lines appearance
}
\examples{
library(ggplot2)
p <- ggplot(mtcars, aes(wt, mpg)) + geom_point()

# Very light minor grid lines
p + adjust_panel_grid_minor(color = "#F5F5F5", linewidth = 0.2)

# Remove minor grid lines
p + adjust_panel_grid_minor(color = NA)
}
