% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ggplot-adjust.R
\name{adjust_plot_margin}
\alias{adjust_plot_margin}
\title{Adjust plot margin}
\usage{
adjust_plot_margin(
  top = NULL,
  right = NULL,
  bottom = NULL,
  left = NULL,
  unit = "pt"
)
}
\arguments{
\item{top}{Top margin. Numeric. NULL to keep default.}

\item{right}{Right margin. Numeric. NULL to keep default.}

\item{bottom}{Bottom margin. Numeric. NULL to keep default.}

\item{left}{Left margin. Numeric. NULL to keep default.}

\item{unit}{Unit for margins. Default "pt". NULL to keep default.}
}
\value{
A ggplot2 theme object.
}
\description{
Adjust plot margin
}
\examples{
library(ggplot2)
p <- ggplot(mtcars, aes(wt, mpg)) + geom_point()

# Add space around plot
p + adjust_plot_margin(top = 20, right = 20, bottom = 20, left = 20)

# Asymmetric margins
p + adjust_plot_margin(top = 30, bottom = 10)
}
