% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ggplot-adjust.R
\name{adjust_strip_background_y}
\alias{adjust_strip_background_y}
\title{Adjust y-direction strip background appearance}
\usage{
adjust_strip_background_y(fill = NULL, color = NULL, linewidth = NULL)
}
\arguments{
\item{fill}{Fill color. NULL to keep default.}

\item{color}{Border color. NULL to keep default.}

\item{linewidth}{Border line width. NULL to keep default.}
}
\value{
A ggplot2 theme object.
}
\description{
Adjust y-direction strip background appearance
}
\examples{
library(ggplot2)
p <- ggplot(mtcars, aes(wt, mpg)) + geom_point() + facet_grid(am ~ cyl)

# Style y-direction strip background
p + adjust_strip_background_y(fill = "#FDE8E8")
}
