% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{apply_missing_values}
\alias{apply_missing_values}
\title{Replace missing codes with NA, handling factor columns}
\usage{
apply_missing_values(col, mis_s, val_s)
}
\arguments{
\item{col}{A vector}

\item{mis_s}{A string with missing value codes}

\item{val_s}{A string with value-label pairs}
}
\value{
A vector with missing values replaced by NA
}
\description{
Replace missing codes with NA, handling factor columns
}
\keyword{internal}
