% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{apply_value_labels}
\alias{apply_value_labels}
\title{Convert column to factor using value-label mapping}
\usage{
apply_value_labels(col, s)
}
\arguments{
\item{col}{A vector}

\item{s}{A string with value-label pairs}
}
\value{
A factor with labeled values
}
\description{
Convert column to factor using value-label mapping
}
\keyword{internal}
