% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\name{axis_order}
\alias{axis_order}
\title{Reorder factor levels on an axis}
\usage{
axis_order(x = NULL, y = NULL)
}
\arguments{
\item{x}{New order of x-axis levels as a character vector. Default: NULL.}

\item{y}{New order of y-axis levels as a character vector. Default: NULL.}
}
\value{
A list of ggplot components.
}
\description{
Reorder factor levels on an axis
}
\examples{
library(ggplot2)

p <- ggplot(mtcars, aes(factor(cyl), mpg)) + geom_boxplot()
# Reverse order
p + axis_order(x = c("8", "6", "4"))
# Custom order
p + axis_order(x = c("6", "4", "8"))
}
\seealso{
\code{\link{axis_x_order}}, \code{\link{axis_y_order}}, \code{\link{axis_text}}
}
