% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\name{axis_x_order}
\alias{axis_x_order}
\title{Reorder x-axis factor levels}
\usage{
axis_x_order(levels = NULL)
}
\arguments{
\item{levels}{New order of x-axis levels as a character vector. Default: NULL.}
}
\value{
A list of ggplot components.
}
\description{
Reorder x-axis factor levels
}
\examples{
library(ggplot2)
p <- ggplot(mtcars, aes(factor(cyl), mpg)) + geom_boxplot()
p + axis_x_order(c("8", "6", "4"))
p + axis_x_order(c("6", "4", "8"))
}
\seealso{
\code{\link{axis_order}}, \code{\link{axis_y_order}}
}
