% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\name{axis_x_text}
\alias{axis_x_text}
\title{Set x-axis tick label text and/or style}
\usage{
axis_x_text(
  text = NULL,
  size = NULL,
  color = NULL,
  angle = NULL,
  face = NULL,
  family = NULL,
  hjust = NULL,
  vjust = NULL
)
}
\arguments{
\item{text}{Named character vector or function for x-axis labels. Names are
original labels, values are new labels. If a function, applied to original
labels. If NULL, removes labels. Default: NULL.}

\item{size}{Font size in points. Default: NULL.}

\item{color}{Text color. Default: NULL.}

\item{angle}{Rotation angle in degrees. Default: NULL.}

\item{face}{Font face. Default: NULL.}

\item{family}{Font family. Default: NULL.}

\item{hjust}{Horizontal justification (0-1). Default: NULL.}

\item{vjust}{Vertical justification (0-1). Default: NULL.}
}
\value{
A list of ggplot components.
}
\description{
Set x-axis tick label text and/or style
}
\examples{
library(ggplot2)
p <- ggplot(mtcars, aes(factor(cyl), mpg)) + geom_boxplot()
p + axis_x_text(c("4" = "4 Cyl", "6" = "6 Cyl", "8" = "8 Cyl"))
p + axis_x_text(c("4" = "Four"))
p + axis_x_text(size = 12, color = "red", angle = 45)
p + axis_x_text(NULL)

# Continuous variable
p2 <- ggplot(mtcars, aes(wt, mpg)) + geom_point()
p2 + axis_x_text(function(x) paste0(x, " tons"))
}
\seealso{
\code{\link{axis_text}}, \code{\link{axis_y_text}}
}
