% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\name{axis_y_order}
\alias{axis_y_order}
\title{Reorder y-axis factor levels}
\usage{
axis_y_order(levels = NULL)
}
\arguments{
\item{levels}{New order of y-axis levels as a character vector. Default: NULL.}
}
\value{
A list of ggplot components.
}
\description{
Reorder y-axis factor levels
}
\examples{
library(ggplot2)
p <- ggplot(mtcars, aes(mpg, factor(cyl))) + geom_boxplot()
p + axis_y_order(c("8", "6", "4"))
p + axis_y_order(c("6", "4", "8"))
}
\seealso{
\code{\link{axis_order}}, \code{\link{axis_x_order}}
}
