% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{clean_sep}
\alias{clean_sep}
\title{Replace Chinese punctuation with ASCII comma}
\usage{
clean_sep(s)
}
\arguments{
\item{s}{A string}
}
\value{
A cleaned string
}
\description{
Replace Chinese punctuation with ASCII comma
}
\keyword{internal}
