% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ggplot.R
\name{clip_off}
\alias{clip_off}
\title{Turn off coordinate clipping}
\usage{
clip_off(clip = "off")
}
\arguments{
\item{clip}{Character, either "on" (default) or "off".}
}
\value{
A coord_cartesian object.
}
\description{
Allows plot elements (like text labels, annotations) to extend
beyond the panel boundaries.
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(wt, mpg)) +
  geom_point() +
  geom_text(aes(label = rownames(mtcars)), hjust = 1.1) +
  theme_sci() +
  clip_off()
}
