% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-colors.R
\name{.check_amount}
\alias{.check_amount}
\title{Check Color Adjustment Amount}
\usage{
.check_amount(amount, range = c(0, 1))
}
\arguments{
\item{amount}{A numeric vector specifying the magnitude of color
adjustment.}

\item{range}{A numeric vector of length two specifying the inclusive
lower and upper bounds allowed for \code{amount}.}
}
\value{
The input \code{amount}, invisibly.
}
\description{
Validate the \code{amount} argument used for color adjustment functions.
}
\keyword{internal}
