% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty-axis.R
\name{.is_discrete_axis}
\alias{.is_discrete_axis}
\title{Check whether a ggplot2 scale is discrete}
\usage{
.is_discrete_axis(scale)
}
\arguments{
\item{scale}{A ggplot2 scale object.}
}
\value{
A single logical value.
}
\description{
Check whether a ggplot2 scale is discrete
}
\keyword{internal}
