% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty-axis.R
\name{.select_discrete_breaks}
\alias{.select_discrete_breaks}
\title{Select approximately n breaks from a discrete scale}
\usage{
.select_discrete_breaks(breaks, n = 5)
}
\arguments{
\item{breaks}{Existing discrete scale breaks.}

\item{n}{Maximum approximate number of breaks.}
}
\value{
A vector containing the selected breaks.
}
\description{
Select approximately n breaks from a discrete scale
}
\keyword{internal}
