% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty-axis.R
\name{ggplot_add.pretty_breaks}
\alias{ggplot_add.pretty_breaks}
\title{Add pretty breaks to a ggplot object}
\usage{
\method{ggplot_add}{pretty_breaks}(object, plot, ...)
}
\arguments{
\item{object}{A \code{pretty_breaks} object.}

\item{plot}{A ggplot object.}

\item{...}{Additional arguments passed by \code{ggplot2::ggplot_add()}.}
}
\value{
A ggplot object with updated x and y scales.
}
\description{
Add pretty breaks to a ggplot object
}
\keyword{internal}
