% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend.R
\name{legend_delete}
\alias{legend_delete}
\title{Remove entire legend or specific aesthetic guides}
\usage{
legend_delete(
  aesthetic = c("all", "color", "fill", "colour", "shape", "size", "alpha", "linetype",
    "group")
)
}
\arguments{
\item{aesthetic}{Aesthetic type: "all", "color", "fill", "colour", "shape",
"size", "alpha", "linetype", or "group". Default "all".}
}
\value{
List of ggplot2 theme/guides components to hide legends.
}
\description{
Remove entire legend or specific aesthetic guides
}
\examples{
library(ggplot2)
p <- ggplot(mtcars, aes(wt, mpg, color = factor(cyl), shape = factor(cyl))) +
  geom_point()

# Remove all legends
p + legend_delete()

# Remove only the color legend
p + legend_delete(aesthetic = "color")

}
