% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend.R
\name{legend_delete_title}
\alias{legend_delete_title}
\title{Remove legend title for specific or all aesthetics}
\usage{
legend_delete_title(
  aesthetic = c("all", "color", "fill", "colour", "shape", "size", "alpha", "linetype",
    "group")
)
}
\arguments{
\item{aesthetic}{Aesthetic type: "all", "color", "fill", "colour", "shape",
"size", "alpha", "linetype", or "group". Default "all".}
}
\value{
List of ggplot2 theme components that blank the legend title.
}
\description{
Remove legend title for specific or all aesthetics
}
\examples{
library(ggplot2)
p <- ggplot(mtcars, aes(wt, mpg, color = factor(cyl))) + geom_point()

# Remove all legend titles
p + legend_delete_title()

# Remove only the color legend title
p + legend_delete_title(aesthetic = "color")

}
