% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend.R
\name{legend_text}
\alias{legend_text}
\title{Relabel legend entries for discrete scales}
\usage{
legend_text(
  labels = NULL,
  aesthetic = c("all", "color", "fill", "colour", "shape", "size", "alpha", "linetype",
    "group")
)
}
\arguments{
\item{labels}{Character vector of new label names, or NULL.}

\item{aesthetic}{Aesthetic type: "all", "color", "fill", "colour", "shape",
"size", "alpha", "linetype", or "group". Default "all".}
}
\value{
List of ggplot2 scale functions with updated labels.
}
\description{
Relabel legend entries for discrete scales
}
\examples{
library(ggplot2)
p <- ggplot(mtcars, aes(wt, mpg, color = factor(cyl))) + geom_point()

# Rename all legend labels
p + legend_text(labels = c("4-Cyl", "6-Cyl", "8-Cyl"))

# Rename only color aesthetic labels
p + legend_text(labels = c("Four", "Six", "Eight"), aesthetic = "color")

}
