% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-colors.R
\name{lighten_colors}
\alias{lighten_colors}
\title{Lighten Colors}
\usage{
lighten_colors(
  colors,
  amount,
  space = c("HCL", "HLS", "combined"),
  plot = FALSE
)
}
\arguments{
\item{colors}{A character vector of colors.}

\item{amount}{A numeric vector specifying the amount by which the colors
should be lightened. Values must be between 0 and 1.}

\item{space}{A character string specifying the color space used for
lightening. One of \code{"HCL"}, \code{"HLS"}, or \code{"combined"}.}

\item{plot}{Logical. If \code{TRUE}, display the resulting colors using
\code{\link[=show_colors]{show_colors()}}. Defaults to \code{FALSE}.}
}
\value{
If \code{plot = FALSE}, a character vector of lightened colors.
If \code{plot = TRUE}, a ggplot object returned by \code{\link[=show_colors]{show_colors()}}.
When \code{amount} contains multiple values, multiple colors are returned
for each input color.
}
\description{
Lighten one or more colors by a specified amount.
}
\details{
This function is a wrapper around \code{\link[colorspace:lighten]{colorspace::lighten()}}. The values of
\code{amount} are applied independently to each color in \code{colors}.
}
\examples{
lighten_colors("#336699", 0.2)

lighten_colors(
  "#336699",
  c(0.1, 0.2, 0.3)
)

lighten_colors(
  "#336699",
  c(0.1, 0.2, 0.3),
  plot = TRUE
)

}
\seealso{
\code{\link[=darken_colors]{darken_colors()}}, \code{\link[=shade_colors]{shade_colors()}}, \code{\link[=show_colors]{show_colors()}}
}
