% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-colors.R
\name{pal}
\alias{pal}
\alias{pal_aaas}
\alias{pal_atlassian}
\alias{pal_bmj}
\alias{pal_bs5}
\alias{pal_cosmic}
\alias{pal_d3}
\alias{pal_flatui}
\alias{pal_frontiers}
\alias{pal_futurama}
\alias{pal_gephi}
\alias{pal_gsea}
\alias{pal_igv}
\alias{pal_iterm}
\alias{pal_jama}
\alias{pal_jco}
\alias{pal_lancet}
\alias{pal_locuszoom}
\alias{pal_material}
\alias{pal_nejm}
\alias{pal_npg}
\alias{pal_observable}
\alias{pal_primer}
\alias{pal_rickandmorty}
\alias{pal_simpsons}
\alias{pal_startrek}
\alias{pal_tron}
\alias{pal_tw3}
\alias{pal_uchicago}
\alias{pal_ucscgb}
\title{Get ggsci palette colors}
\usage{
pal(palette = "jama", n = NULL, alpha = 1, amount = NULL, show = FALSE)

pal_aaas(n = NULL, alpha = 1, amount = NULL, show = FALSE)

pal_atlassian(n = NULL, alpha = 1, amount = NULL, show = FALSE)

pal_bmj(n = NULL, alpha = 1, amount = NULL, show = FALSE)

pal_bs5(n = NULL, alpha = 1, amount = NULL, show = FALSE)

pal_cosmic(n = NULL, alpha = 1, amount = NULL, show = FALSE)

pal_d3(n = NULL, alpha = 1, amount = NULL, show = FALSE)

pal_flatui(n = NULL, alpha = 1, amount = NULL, show = FALSE)

pal_frontiers(n = NULL, alpha = 1, amount = NULL, show = FALSE)

pal_futurama(n = NULL, alpha = 1, amount = NULL, show = FALSE)

pal_gephi(n = NULL, alpha = 1, amount = NULL, show = FALSE)

pal_gsea(n = NULL, alpha = 1, amount = NULL, show = FALSE)

pal_igv(n = NULL, alpha = 1, amount = NULL, show = FALSE)

pal_iterm(n = NULL, alpha = 1, amount = NULL, show = FALSE)

pal_jama(n = NULL, alpha = 1, amount = NULL, show = FALSE)

pal_jco(n = NULL, alpha = 1, amount = NULL, show = FALSE)

pal_lancet(n = NULL, alpha = 1, amount = NULL, show = FALSE)

pal_locuszoom(n = NULL, alpha = 1, amount = NULL, show = FALSE)

pal_material(n = NULL, alpha = 1, amount = NULL, show = FALSE)

pal_nejm(n = NULL, alpha = 1, amount = NULL, show = FALSE)

pal_npg(n = NULL, alpha = 1, amount = NULL, show = FALSE)

pal_observable(n = NULL, alpha = 1, amount = NULL, show = FALSE)

pal_primer(n = NULL, alpha = 1, amount = NULL, show = FALSE)

pal_rickandmorty(n = NULL, alpha = 1, amount = NULL, show = FALSE)

pal_simpsons(n = NULL, alpha = 1, amount = NULL, show = FALSE)

pal_startrek(n = NULL, alpha = 1, amount = NULL, show = FALSE)

pal_tron(n = NULL, alpha = 1, amount = NULL, show = FALSE)

pal_tw3(n = NULL, alpha = 1, amount = NULL, show = FALSE)

pal_uchicago(n = NULL, alpha = 1, amount = NULL, show = FALSE)

pal_ucscgb(n = NULL, alpha = 1, amount = NULL, show = FALSE)
}
\arguments{
\item{palette}{Palette name. Available options:
\describe{
\item{aaas}{AAAS journal color palettes}
\item{atlassian}{Atlassian Design System palette}
\item{bmj}{BMJ color palettes}
\item{bs5}{Bootstrap 5 color palettes}
\item{cosmic}{COSMIC color palettes}
\item{d3}{D3.js color palettes}
\item{flatui}{Flat UI color palettes}
\item{frontiers}{Frontiers journal color palettes}
\item{futurama}{Futurama color palettes}
\item{gephi}{Gephi color palettes}
\item{gsea}{The GSEA GenePattern color palettes}
\item{igv}{Integrative Genomics Viewer (IGV) color palettes}
\item{iterm}{iTerm color palettes}
\item{jama}{Journal of the American Medical Association color palettes}
\item{jco}{Journal of Clinical Oncology color palettes}
\item{lancet}{Lancet journal color palettes}
\item{locuszoom}{LocusZoom color palette}
\item{material}{Material Design color palettes}
\item{nejm}{NEJM color palettes}
\item{npg}{NPG journal color palettes}
\item{observable}{Observable 10 color palette}
\item{primer}{Primer design system palette}
\item{rickandmorty}{Rick and Morty color palettes}
\item{simpsons}{The Simpsons color palettes}
\item{startrek}{Star Trek color palettes}
\item{tron}{Tron Legacy color palettes}
\item{tw3}{Tailwind CSS color palettes}
\item{uchicago}{The University of Chicago color palettes}
\item{ucscgb}{UCSC Genome Browser color palette}
}}

\item{n}{Number of colors. Defaults to the maximum number of the palette if NULL}

\item{alpha}{Transparency level, 0 (transparent) to 1 (opaque)}

\item{amount}{Numeric vector for lightening/darkening colors.
Negative values lighten, positive values darken. Values must be between -1 and 1.
Passed to \code{\link{shade_colors}}.}

\item{show}{Whether to display the color}
}
\description{
Unified interface to get colors from various ggsci palettes.
This is the core function called by all pal_*() wrapper functions.
}
\examples{
pal("jama")
pal("npg", 3)
pal("lancet", alpha = 0.5, show = TRUE)
pal("nejm", 5, amount = c(-0.3, 0, 0.3))
}
