% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{parse_missing_values}
\alias{parse_missing_values}
\title{Parse "999, 998" into numeric vector of missing codes}
\usage{
parse_missing_values(s)
}
\arguments{
\item{s}{A string with missing value codes}
}
\value{
A numeric vector
}
\description{
Parse "999, 998" into numeric vector of missing codes
}
\keyword{internal}
