% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{parse_value_pairs}
\alias{parse_value_pairs}
\title{Parse "1=Male, 2=Female" into numeric values and character labels}
\usage{
parse_value_pairs(s)
}
\arguments{
\item{s}{A string with value-label pairs}
}
\value{
A list with values (numeric) and labels (character)
}
\description{
Parse "1=Male, 2=Female" into numeric values and character labels
}
\keyword{internal}
