% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{read_codes}
\alias{read_codes}
\title{Read codebook from file or validate data frame}
\usage{
read_codes(codes)
}
\arguments{
\item{codes}{A data.frame or file path (.xlsx, .xls, .csv)}
}
\value{
A data.frame containing the codebook
}
\description{
Read codebook from file or validate data frame
}
\keyword{internal}
