% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ggplot-remove.R
\name{remove_all}
\alias{remove_all}
\title{Create a completely blank theme}
\usage{
remove_all()
}
\value{
A ggplot2 theme object with everything removed.
}
\description{
Removes all theme elements including axes, legends, strips, panels,
backgrounds, titles, etc. Only the data remains visible.
}
\examples{
library(ggplot2)
p <- ggplot(mtcars, aes(wt, mpg)) + geom_point()
p + remove_all()
}
