% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ggplot-remove.R
\name{remove_axis}
\alias{remove_axis}
\title{Remove both axes completely}
\usage{
remove_axis()
}
\value{
A ggplot2 theme object with both axes removed.
}
\description{
Remove both axes completely
}
\examples{
library(ggplot2)
p <- ggplot(mtcars, aes(wt, mpg)) + geom_point()
p + remove_axis()
}
