% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ggplot-remove.R
\name{remove_axis_line}
\alias{remove_axis_line}
\title{Remove both axis lines}
\usage{
remove_axis_line()
}
\value{
A ggplot2 theme object with both axis lines removed.
}
\description{
Remove both axis lines
}
\examples{
library(ggplot2)
p <- ggplot(mtcars, aes(wt, mpg)) + geom_point()
p + remove_axis_line()
}
