% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ggplot-remove.R
\name{remove_axis_line_x}
\alias{remove_axis_line_x}
\title{Remove x-axis line}
\usage{
remove_axis_line_x()
}
\value{
A ggplot2 theme object with x-axis line removed.
}
\description{
Remove x-axis line
}
\examples{
library(ggplot2)
p <- ggplot(mtcars, aes(wt, mpg)) + geom_point()
p + remove_axis_line_x()
}
