% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ggplot-remove.R
\name{remove_axis_text}
\alias{remove_axis_text}
\title{Remove both axis text (tick labels)}
\usage{
remove_axis_text()
}
\value{
A ggplot2 theme object with both axis text removed.
}
\description{
Remove both axis text (tick labels)
}
\examples{
library(ggplot2)
p <- ggplot(mtcars, aes(wt, mpg)) + geom_point()
p + remove_axis_text()
}
