% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ggplot-remove.R
\name{remove_axis_text_x}
\alias{remove_axis_text_x}
\title{Remove x-axis text (tick labels)}
\usage{
remove_axis_text_x()
}
\value{
A ggplot2 theme object with x-axis text removed.
}
\description{
Remove x-axis text (tick labels)
}
\examples{
library(ggplot2)
p <- ggplot(mtcars, aes(wt, mpg)) + geom_point()
p + remove_axis_text_x()
}
