% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ggplot-remove.R
\name{remove_axis_text_y}
\alias{remove_axis_text_y}
\title{Remove y-axis text (tick labels)}
\usage{
remove_axis_text_y()
}
\value{
A ggplot2 theme object with y-axis text removed.
}
\description{
Remove y-axis text (tick labels)
}
\examples{
library(ggplot2)
p <- ggplot(mtcars, aes(wt, mpg)) + geom_point()
p + remove_axis_text_y()
}
