% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ggplot-remove.R
\name{remove_axis_ticks}
\alias{remove_axis_ticks}
\title{Remove both axis ticks}
\usage{
remove_axis_ticks()
}
\value{
A ggplot2 theme object with both axis ticks removed.
}
\description{
Remove both axis ticks
}
\examples{
library(ggplot2)
p <- ggplot(mtcars, aes(wt, mpg)) + geom_point()
p + remove_axis_ticks()
}
