% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ggplot-remove.R
\name{remove_axis_ticks_y}
\alias{remove_axis_ticks_y}
\title{Remove y-axis ticks}
\usage{
remove_axis_ticks_y()
}
\value{
A ggplot2 theme object with y-axis ticks removed.
}
\description{
Remove y-axis ticks
}
\examples{
library(ggplot2)
p <- ggplot(mtcars, aes(wt, mpg)) + geom_point()
p + remove_axis_ticks_y()
}
