% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ggplot-remove.R
\name{remove_axis_title}
\alias{remove_axis_title}
\title{Remove both axis titles}
\usage{
remove_axis_title()
}
\value{
A ggplot2 theme object with both axis titles removed.
}
\description{
Remove both axis titles
}
\examples{
library(ggplot2)
p <- ggplot(mtcars, aes(wt, mpg)) + geom_point()
p + remove_axis_title()
}
