% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ggplot-remove.R
\name{remove_axis_y}
\alias{remove_axis_y}
\title{Remove entire y-axis (title, text, ticks, line)}
\usage{
remove_axis_y()
}
\value{
A ggplot2 theme object with entire y-axis removed.
}
\description{
Remove entire y-axis (title, text, ticks, line)
}
\examples{
library(ggplot2)
p <- ggplot(mtcars, aes(wt, mpg)) + geom_point()
p + remove_axis_y()
}
