% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ggplot-remove.R
\name{remove_caption}
\alias{remove_caption}
\title{Remove caption}
\usage{
remove_caption()
}
\value{
A ggplot2 theme object with caption removed.
}
\description{
Remove caption
}
\examples{
library(ggplot2)
p <- ggplot(mtcars, aes(wt, mpg)) + geom_point() + labs(caption = "Caption")
p + remove_caption()
}
