% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ggplot-remove.R
\name{remove_grid}
\alias{remove_grid}
\title{Remove all grid lines (major and minor)}
\usage{
remove_grid()
}
\value{
A ggplot2 theme object with all grid lines removed.
}
\description{
Remove all grid lines (major and minor)
}
\examples{
library(ggplot2)
p <- ggplot(mtcars, aes(wt, mpg)) + geom_point()
p + remove_grid()
}
