% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ggplot-remove.R
\name{remove_grid_major}
\alias{remove_grid_major}
\title{Remove major grid lines}
\usage{
remove_grid_major()
}
\value{
A ggplot2 theme object with major grid lines removed.
}
\description{
Remove major grid lines
}
\examples{
library(ggplot2)
p <- ggplot(mtcars, aes(wt, mpg)) + geom_point()
p + remove_grid_major()
}
