% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ggplot-remove.R
\name{remove_grid_major_y}
\alias{remove_grid_major_y}
\title{Remove y-direction major grid lines}
\usage{
remove_grid_major_y()
}
\value{
A ggplot2 theme object with y-direction major grid lines removed.
}
\description{
Remove y-direction major grid lines
}
\examples{
library(ggplot2)
p <- ggplot(mtcars, aes(wt, mpg)) + geom_point()
p + remove_grid_major_y()
}
