% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ggplot-remove.R
\name{remove_grid_minor}
\alias{remove_grid_minor}
\title{Remove minor grid lines}
\usage{
remove_grid_minor()
}
\value{
A ggplot2 theme object with minor grid lines removed.
}
\description{
Remove minor grid lines
}
\examples{
library(ggplot2)
p <- ggplot(mtcars, aes(wt, mpg)) + geom_point()
p + remove_grid_minor()
}
