% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ggplot-remove.R
\name{remove_legend}
\alias{remove_legend}
\title{Remove entire legend}
\usage{
remove_legend()
}
\value{
A ggplot2 theme object with legend removed.
}
\description{
Remove entire legend
}
\examples{
library(ggplot2)
p <- ggplot(mtcars, aes(wt, mpg, color = factor(cyl))) + geom_point()
p + remove_legend()
}
