% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ggplot-remove.R
\name{remove_legend_box}
\alias{remove_legend_box}
\title{Remove legend box background}
\usage{
remove_legend_box()
}
\value{
A ggplot2 theme object with legend box background removed.
}
\description{
Remove legend box background
}
\examples{
library(ggplot2)
p <- ggplot(mtcars, aes(wt, mpg, color = factor(cyl))) + geom_point()
p + remove_legend_box()
}
