% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ggplot-remove.R
\name{remove_legend_key}
\alias{remove_legend_key}
\title{Remove legend key background}
\usage{
remove_legend_key()
}
\value{
A ggplot2 theme object with legend key removed.
}
\description{
Remove legend key background
}
\examples{
library(ggplot2)
p <- ggplot(mtcars, aes(wt, mpg, color = factor(cyl))) + geom_point()
p + remove_legend_key()
}
