% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ggplot-remove.R
\name{remove_legend_text}
\alias{remove_legend_text}
\title{Remove legend text}
\usage{
remove_legend_text()
}
\value{
A ggplot2 theme object with legend text removed.
}
\description{
Remove legend text
}
\examples{
library(ggplot2)
p <- ggplot(mtcars, aes(wt, mpg, color = factor(cyl))) + geom_point()
p + remove_legend_text()
}
