% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ggplot-remove.R
\name{remove_legend_title}
\alias{remove_legend_title}
\title{Remove legend title}
\usage{
remove_legend_title()
}
\value{
A ggplot2 theme object with legend title removed.
}
\description{
Remove legend title
}
\examples{
library(ggplot2)
p <- ggplot(mtcars, aes(wt, mpg, color = factor(cyl))) + geom_point()
p + remove_legend_title()
}
