% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ggplot-remove.R
\name{remove_panel}
\alias{remove_panel}
\title{Remove entire panel (background, border, grid)}
\usage{
remove_panel()
}
\value{
A ggplot2 theme object with entire panel decorations removed.
}
\description{
Remove entire panel (background, border, grid)
}
\examples{
library(ggplot2)
p <- ggplot(mtcars, aes(wt, mpg)) + geom_point()
p + remove_panel()
}
