% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ggplot-remove.R
\name{remove_panel_background}
\alias{remove_panel_background}
\title{Remove panel background}
\usage{
remove_panel_background()
}
\value{
A ggplot2 theme object with panel background removed.
}
\description{
Remove panel background
}
\examples{
library(ggplot2)
p <- ggplot(mtcars, aes(wt, mpg)) + geom_point()
p + remove_panel_background()
}
