% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ggplot-remove.R
\name{remove_panel_border}
\alias{remove_panel_border}
\title{Remove panel border}
\usage{
remove_panel_border()
}
\value{
A ggplot2 theme object with panel border removed.
}
\description{
Remove panel border
}
\examples{
library(ggplot2)
p <- ggplot(mtcars, aes(wt, mpg)) + geom_point()
p + remove_panel_border()
}
