% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ggplot-remove.R
\name{remove_plot_background}
\alias{remove_plot_background}
\title{Remove plot background}
\usage{
remove_plot_background()
}
\value{
A ggplot2 theme object with plot background removed.
}
\description{
Remove plot background
}
\examples{
library(ggplot2)
p <- ggplot(mtcars, aes(wt, mpg)) + geom_point()
p + remove_plot_background()
}
