% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ggplot-remove.R
\name{remove_strip}
\alias{remove_strip}
\title{Remove entire strip (text and background)}
\usage{
remove_strip()
}
\value{
A ggplot2 theme object with entire strip removed.
}
\description{
Remove entire strip (text and background)
}
\examples{
library(ggplot2)
p <- ggplot(mtcars, aes(wt, mpg)) + geom_point() + facet_wrap(~ cyl)
p + remove_strip()
}
