% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ggplot-remove.R
\name{remove_strip_background}
\alias{remove_strip_background}
\title{Remove strip background}
\usage{
remove_strip_background()
}
\value{
A ggplot2 theme object with strip background removed.
}
\description{
Remove strip background
}
\examples{
library(ggplot2)
p <- ggplot(mtcars, aes(wt, mpg)) + geom_point() + facet_wrap(~ cyl)
p + remove_strip_background()
}
