% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ggplot-remove.R
\name{remove_strip_background_y}
\alias{remove_strip_background_y}
\title{Remove y-direction strip background}
\usage{
remove_strip_background_y()
}
\value{
A ggplot2 theme object with y-direction strip background removed.
}
\description{
Remove y-direction strip background
}
\examples{
library(ggplot2)
p <- ggplot(mtcars, aes(wt, mpg)) + geom_point() + facet_grid(am ~ cyl)
p + remove_strip_background_y()
}
