% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ggplot-remove.R
\name{remove_strip_text}
\alias{remove_strip_text}
\title{Remove strip text (facet labels)}
\usage{
remove_strip_text()
}
\value{
A ggplot2 theme object with strip text removed.
}
\description{
Remove strip text (facet labels)
}
\examples{
library(ggplot2)
p <- ggplot(mtcars, aes(wt, mpg)) + geom_point() + facet_wrap(~ cyl)
p + remove_strip_text()
}
