% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ggplot-remove.R
\name{remove_subtitle}
\alias{remove_subtitle}
\title{Remove subtitle}
\usage{
remove_subtitle()
}
\value{
A ggplot2 theme object with subtitle removed.
}
\description{
Remove subtitle
}
\examples{
library(ggplot2)
p <- ggplot(mtcars, aes(wt, mpg)) + geom_point() + labs(subtitle = "Subtitle")
p + remove_subtitle()
}
