% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ggplot-remove.R
\name{remove_tag}
\alias{remove_tag}
\title{Remove tag}
\usage{
remove_tag()
}
\value{
A ggplot2 theme object with tag removed.
}
\description{
Remove tag
}
\examples{
library(ggplot2)
p <- ggplot(mtcars, aes(wt, mpg)) + geom_point() + labs(tag = "A")
p + remove_tag()
}
