% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ggplot-remove.R
\name{remove_title}
\alias{remove_title}
\title{Remove main title}
\usage{
remove_title()
}
\value{
A ggplot2 theme object with title removed.
}
\description{
Remove main title
}
\examples{
library(ggplot2)
p <- ggplot(mtcars, aes(wt, mpg)) + geom_point() + ggtitle("My Title")
p + remove_title()
}
