% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ggplot.R
\name{rotate_x_axis}
\alias{rotate_x_axis}
\title{Rotate x-axis labels}
\usage{
rotate_x_axis(angle = 45, hjust = NULL, vjust = NULL)
}
\arguments{
\item{angle}{Rotation angle in degrees.}

\item{hjust}{Horizontal justification.}

\item{vjust}{Vertical justification.}
}
\value{
A ggplot2 theme object.
}
\description{
Rotate x-axis tick labels.
}
\examples{
ggplot2::ggplot(mtcars, ggplot2::aes(factor(cyl), mpg)) +
  ggplot2::geom_point() +
  rotate_x_axis()

ggplot2::ggplot(mtcars, ggplot2::aes(factor(cyl), mpg)) +
  ggplot2::geom_point() +
  rotate_x_axis(angle = 90)
}
