% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ggplot.R
\name{rotate_y_axis}
\alias{rotate_y_axis}
\title{Rotate y-axis labels}
\usage{
rotate_y_axis(angle = 45, hjust = NULL, vjust = NULL)
}
\arguments{
\item{angle}{Rotation angle in degrees.}

\item{hjust}{Horizontal justification.}

\item{vjust}{Vertical justification.}
}
\value{
A ggplot2 theme object.
}
\description{
Rotate y-axis tick labels.
}
\examples{
ggplot2::ggplot(mtcars, ggplot2::aes(mpg, factor(cyl))) +
  ggplot2::geom_point() +
  rotate_y_axis()

ggplot2::ggplot(mtcars, ggplot2::aes(mpg, factor(cyl))) +
  ggplot2::geom_point() +
  rotate_y_axis(angle = 45)
}
