% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-colors.R
\name{scale_color_jama}
\alias{scale_color_jama}
\alias{scale_fill_jama}
\title{JAMA color scale for ggplot2}
\usage{
scale_color_jama(alpha = 1, amount = NULL, ...)

scale_fill_jama(alpha = 1, amount = NULL, ...)
}
\arguments{
\item{alpha}{Transparency level, 0 (transparent) to 1 (opaque)}

\item{amount}{Numeric vector for lightening/darkening colors.
Negative values lighten, positive values darken. Values must be between -1 and 1.
Passed to \code{\link{shade_colors}}.}

\item{...}{Additional arguments passed to \code{\link[ggplot2]{discrete_scale}}}
}
\description{
JAMA color scale for ggplot2
}
