% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-colors.R
\name{scale_color_observable}
\alias{scale_color_observable}
\alias{scale_fill_observable}
\title{Observable color scale for ggplot2}
\usage{
scale_color_observable(alpha = 1, amount = NULL, ...)

scale_fill_observable(alpha = 1, amount = NULL, ...)
}
\arguments{
\item{alpha}{Transparency level, 0 (transparent) to 1 (opaque)}

\item{amount}{Numeric vector for lightening/darkening colors.}

\item{...}{Additional arguments passed to \code{\link[ggplot2]{discrete_scale}}}
}
\description{
Observable color scale for ggplot2
}
