% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-sci.R
\name{set_family}
\alias{set_family}
\title{Can be added directly to a ggplot with `+` operator.}
\usage{
set_family(family = "sans")
}
\arguments{
\item{family}{Font family name. e.g., "serif", "sans", "Arial", "Times New Roman"}
}
\value{
A ggplot2 theme object
}
\description{
Can be added directly to a ggplot with `+` operator.
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(x = wt, y = mpg)) +
  geom_point() +
  set_family("serif")
}
