% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{set_label}
\alias{set_label}
\title{Attach label attribute to a column}
\usage{
set_label(col, s)
}
\arguments{
\item{col}{A vector}

\item{s}{A string for the label}
}
\value{
A vector with label attribute set
}
\description{
Attach label attribute to a column
}
\keyword{internal}
