% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-colors.R
\name{show_colors}
\alias{show_colors}
\title{Display a Color Palette}
\usage{
show_colors(colors, labels = TRUE, size = 3.5)
}
\arguments{
\item{colors}{A character vector of colors.}

\item{labels}{Logical. Whether to display color codes below each color.
Defaults to \code{TRUE}.}

\item{size}{Numeric. The text size of color labels. Defaults to \code{3.5}.}
}
\value{
A ggplot object displaying the supplied colors.
}
\description{
Display a vector of colors as a color palette using ggplot2.
}
\examples{
show_colors(c("#336699", "#CC3333", "#66CC99"))

show_colors(
  c("#336699", "#CC3333", "#66CC99"),
  labels = FALSE
)

}
